Change #6: Prototype-ish
- Finish button code, for REAL. mousePressed seemed to be working, and
against all reasonable judgement calls I could have made, I chose to
keep it. Wrong choice. Changed to mouseReleased via mouseIsReleased
boolean, constantly being set to false in draw except for when
mouseIsReleased, at which point it is true for a frame, at which
point—if mouse is over a button—the screen will change.
- Fixed button drawing to be more uniform and, well, rational. Button
text is now center aligned, and _buttonWidth is now defined by
textWidth of _labelText.
- Minor text edits, as well as slow removal of Placeholder names (some
are now named, and are properly linked)
- Added comments!
- Improved name replacement somewhat (displaying the player’s inputted
text is still not working smoothly, though the naming is). Added
failsafe on non-naming by way of the names Zack, Abhay, and Rana for
main, friend, and enemy (yes, enemy is female, and yes, Rana is a male
name. Humor). If player chooses not to name for some reason, these
names are automatically input as the chosen names.
- Removed clutter (useless functions, commented out sections…)