Change #8: Music To My Ears
- Implemented file selection via JFileChooser. Allows user to select
any .mp3, .wav, or .m4a (not sure if .m4a files are functioning as of
yet) and will play that song.
- Created alternate Button constructor to interface with music
selection rather than screen switching
- Slightly altered name selection process: if name is written AND enter
is pressed, name is set, else name will not be set. This does not draw
the name as of yet (planned for future)
- Implemented println via case
- Changed textadventure.zk to SchoolQuest.zk

Current Bugs:

- File selector is buggy as all HECK.
- Clicking before project is loaded will sometimes freeze it
- Quitting out of file selection without choosing a song will sometimes
crash program (though measures were put in place to prevent this)

Current Objectives:

- Name drawing
- Inventory
- Work. Harder. Damn. It.
- Write
- Polish