Change #11: Taking Inventory
- Implemented “inventory” via Dialog. Not yet implemented into text
file (baby steps, baby steps), but checks items in inventory ArrayList,
and adds it to Dialog.
- Inventory button repurposed to pull up drop-down “inventory” menu.
Certain screen trigger flags which add items to the menu (as a test,
Screen 1 automatically adds computer to inventory). Different cases can
be triggered depending on the screen an item is used, with the default
case being a reference to the “there’s a time and place for everything,
but not now” line from Pokémon if an item is used at an unexpected
time. With further writing, this inventory can be further expanded upon.
- Converted StringLists to ArrayList<Strings>, for easier usage.
- Played around with locker number randomization: if @PlayerName &
@FriendName are ≥ 20, the two will be annoyed and split up, triggering
screen 40 (?), whereas if they are under 20 apart they will remain
together and trigger screen 41 (?). These numbers are subject to change.
- Added booleans for schoolQuest and theVoid respectively. Certain
actions will only trigger if in the right game.
- Removed “Inventory Screen”
- Removed all CheckPresses/Update function under Button class except
main. Now using cases to check Button type (“Story” triggers _goesTo =
currentScreen, “Music” triggers JFileChooser, “Inventory” JOpenDialog,
“Menu” state = _goesTo)
- Added a few screens in SchoolQuest.zk (fewer than I’d hoped)
- Generally fell in love with switch/case. Will attempt to further cut
out unnecessary if statements.

To Finish:

- Writing/Screens
- Non-linear inventory (effects of encountering items in non-set order)
- README
- Fix crashing issue with JFileChooser