Change #12: Voided
- Set up Void’s paths, allowing it to be mostly playable (Inventory
functionality is broken on it for the time being, beyond simply being
unimplemented; inventory from SchoolQuest carries over to it if it is
selected after)
- Added some more switch/cases: currentScreen switches in keyPressed()
and CheckBooleans()
- Added boolean functions computerObtained() and robotObtained()
(placeholder items), in order to allow checking inventory for items
- Mostly finished inventory functionality for SchoolQuest. Added choice
== null checking (forgotten when porting inventory checker over from
proof of concept sketch), as well as currentScreen switch under
individual inventory choice cases (i.e. computer on screen 2 will
prompt user for password, if password correct take user to 40), else it
will default to saying cannot be used at this time. As of now, entirety
of inventory items are placeholders. Planning to continue working on
inventory once more of SchoolQuest screens are completed.
- Cut some unnecessary code out of FileSelector.
- Cut down on Parse() code, now starting character is used as a switch
in order to add screens rather than reusing the entirety of the body
code with no difference but a different starting character.